requirements.txt
Flask
google-generativeai
google-ai-generativelanguage
python-dotenv
elevenlabs
requests
uuid
